requirements.txt
django >=1.9, <2
pyephem==3.7.6.0
djangorestframework==3.6.2
spacetrack==0.12.0
